feat: Add stack option to project creation
This commit adds a new `Stack` field to the `New.Parsed` and `New.Opts` types in the `Commands.fs` and `Types.fs` files respectively. The stack option allows users to specify a custom stack name when creating a new project. If no stack name is provided, it defaults to "pinkdiamond".