Most of velero (#2005)
* Adding velero

* Fix minio

* feat: Add new buckets to Minio configuration

This commit adds two new buckets, 'thecluster-s3' and 'velero', to the Minio configuration in order to support additional functionality.

* Remove unnecessary mapping in vcs.xml

The commit removes an unnecessary mapping in the vcs.xml file. The mapping for the directory "$PROJECT_DIR$/submodules/cert-manager" has been removed.

* Update Minio resource limits and remove unnecessary bucket configurations

- Update CPU limit from '100m' to '500m'
- Update memory limit from '256Mi' to '512Mi'
- Remove unnecessary bucket configurations for 'velero'

This commit updates the resource limits for Minio, increasing the CPU and memory limits. Additionally, it removes the unnecessary bucket configuration for 'velero'.

* feat: Update Minio TLS configuration

- Updated the issuer reference to use the stage cluster issuer instead of the root cluster issuer.
- Added additional DNS names for Minio and console endpoints.
- Modified the buckets array to include a single bucket name.
- Updated annotations for SSL redirect and cluster issuer in the ingress configuration.

This commit improves the TLS configuration for Minio, adding support for multiple DNS names and using the appropriate cluster issuers.

* feat: Add velero namespace with custom labels

This commit adds a new namespace called "velero" to the shared index file. The namespace includes custom labels, specifically setting the label "thecluster.io/trust" to "root". This change allows for more granular control and organization within the infrastructure.

* feat: Add Velero installation for THECLUSTER

This commit adds the necessary configuration and resources to install Velero in THECLUSTER. The following significant changes were made:

- Added a new file `Pulumi.pinkdiamond.yaml` with configuration settings for Minio and Backblaze
- Updated `Pulumi.yaml` to include versions of dependencies (bitnamiKubectl, velero, veleroCsi, veleroAws)
- Updated `config.ts` to include a new interface `Backblaze` and added required properties for Backblaze credentials
- Updated `index.ts` to create an S3 bucket for backups, configure secrets for Minio and Backblaze credentials, and deploy Velero using Helm
- Updated `package-lock.json` with new dependencies

These changes enable the installation of Velero in THECLUSTER with support for both Minio and Backblaze as backup storage locations.

* Update commonName in minio TLS certificate and add additional DNS names for s3.thecluster.io. Update OAUTH2_PROXY_UPSTREAMS value to use port 443 instead of a variable.