feat: Add support for custom command line options
This commit adds support for custom command line options in the `Args.fs` file. Specifically, it introduces two new types: `Language` and `ProjectType`, which define alternative names for certain command line arguments. However, it should be noted that the alternative names are not currently functional.

Additionally, this commit updates the usage of these types in other files such as `Commands.fs`, `Domain.fs`, and `Program.fs`. The changes include renaming the type `New` to `NewProject` in the `Commands.fs` file and updating function signatures accordingly.