feat: Add dynamic redirect URIs for keycloak clients
This commit adds the ability to dynamically generate valid redirect URIs for keycloak clients in two files. The code now imports the `redirectUris` function from `@unstoppablemango/thecluster/apps/keycloak` and uses it to set the `validRedirectUris` property of the keycloak client objects. This allows for more flexibility and easier management of redirect URIs.

Changes:
- Import `redirectUris` function from `@unstoppablemango/thecluster/apps/keycloak`
- Use `redirectUris` function to set `validRedirectUris` property of keycloak client objects in two files