Bunch of kubebuilder tests (#2880)
* Durdlin

* Updated Makefile and added Pulumi version

- Introduced a new file to specify the Pulumi version
- Updated paths for Pulumi and Ginkgo in the Makefile
- Added 'ensure' target to ensure necessary binaries are present
- Added 'clean' target to remove all generated binaries
- Included steps to install Ginkgo and Pulumi in their respective directories

* Updated Makefile for improved testing

The Makefile has been updated to enhance the testing process. Changes include:
- Modified TEST_SENTINELS and GINKGO_REPORTS variables for better report generation
- Adjusted 'testf' target to clean tests before running
- Updated 'gen' target to use GO_GEN_SRC variable
- Added JSON reports with separate keep option in TEST_FLAGS
- Simplified the creation of TEST_SENTINELS by removing unnecessary echo commands
- Introduced a new target '.make/clean_tests' to remove test sentinels

* Updated Makefile for test dependencies

The Makefile has been updated to include new dependencies for the tests. The TEST_SENTINELS now depend on GO_SRC and bin/ginkgo. Additionally, a new dependency has been added specifically for cmd/kubebuilder/TEST_REPORT, which depends on bin/kubebuilder and bin/kubectl.

* Refined test dependencies in Makefile

The Makefile has been updated to better specify the dependencies for each test report. This change ensures that only relevant source files trigger a re-run of tests, improving efficiency. Additionally, an end-to-end test was added to verify the creation of the PROJECT file.

* Refactor Makefile for test execution

Simplified the dependencies in the Makefile for running tests. Removed specific file filters from test report targets and replaced them with a more general filter. Also, adjusted the directory sorting logic to exclude 'theclusterv1alpha1' during test execution.

* Updated E2E test and directory check

Enhanced the E2E test by adding a new context that checks for the existence of a Makefile. Also, updated the directory check function to include 'Makefile' in the list of allowed files. This ensures that projects with a Makefile are not flagged as erroneous.

* Expanded list of allowed files in init function

The init function has been updated to include a wider range of file types that it can handle without modification. This includes various configuration and project files, such as .dockerignore, .editorconfig, buf.gen.yaml, global.json and more. The tests have also been adjusted accordingly to ensure these new file types are handled correctly.