Added stop function and version control to TestCluster
In this update, we've introduced a new stop function to the TestCluster struct in the utils package. This allows for graceful shutdown of clusters. We've also added a way to specify the k3s version when creating a new TestCluster instance. Additionally, we have included default deadline variable for stopping the cluster.

On top of these changes, we've created new tests in cluster_test.go to ensure that our new features are working as expected. These tests cover creation of a new TestCluster, setting k3s version and stopping the cluster successfully even if it hasn't been started yet.