Updated test suite labels and status
In this update, we've made changes to the labeling of our test suites. The "Kubebuilder Suite" and "e2e suite" now have an additional label "E2E". This will help in better categorization and identification of these tests. Additionally, the status of the 'controller' test in operator_e2e_test has been marked as 'Pending'.