Added Pulumi stack initialization status
Introduced a new field to track the initialization status of an App's Pulumi stack. This is represented as a condition in the App's status, which is set to false by default indicating that the stack has not been initialized yet. Corresponding tests have also been added to verify this behavior.