Actually operating on the operator (#2899)
* Updated AppSpec and kubebuilder script

In the operator API, removed scaffolding comments and replaced 'Foo' field with 'Repository' in AppSpec. In the kubebuilder script, added a prompt for user input of 'Group', replacing hardcoded value.

* Go crap

* Updated Makefile for better project management

The Makefile has been updated to improve the build process and project management. A new 'manifests' target has been added to generate controller and operator manifests. The 'tidy' target now includes a step to tidy up the Go modules in the operator directory. This will help keep our dependencies clean and up-to-date, ensuring smoother builds and less clutter in our codebase.

* Fleshing out the App resource

* Updated scaffolding options in AppSpec

Enhanced the 'scaffold' field in the AppSpec struct to support two types of scaffolding: "typescript" and "helm". The default option is "typescript", which performs scaffolding for TypeScript. The new option, "helm", creates an empty helm chart in the app path. These changes are reflected in both the Go code and YAML configuration file.

* Implementing the thing

* Added new rule to .gitignore

A new ignore rule has been added to the .gitignore file. This will prevent end-to-end test folders from being tracked by Git, improving repository cleanliness and reducing unnecessary files in version control.

* Updated 'Manage' field to pointer type and added management logic

The 'Manage' field in the AppSpec struct has been updated from a boolean to a pointer to a boolean. This allows for three states: true, false, and nil. The management logic in the app controller has been updated accordingly. If 'Manage' is set to nil or true, the app is managed by default. If it's set to false, the app is not managed. A utility function for creating a bool pointer was also added. Tests have been updated and expanded to cover these changes.

* App conditions

* Added Pulumi stack initialization status

Introduced a new field to track the initialization status of an App's Pulumi stack. This is represented as a condition in the App's status, which is set to false by default indicating that the stack has not been initialized yet. Corresponding tests have also been added to verify this behavior.

* Enabled E2E tests in CI workflow

The End-to-End (E2E) tests that were previously commented out have been enabled again in the continuous integration (CI) workflow. This change will ensure a more comprehensive testing process before deployment.