Start to decrease the scope of these shenanigans (#2903)
* Make crap again

* Added TypeScript templates for Pulumi

This update introduces new TypeScript templates for Pulumi. The changes include:
- A .helmignore file to specify patterns to ignore when building packages
- Chart.yaml, Pulumi.yaml, config.ts, index.ts files with basic configurations and settings
- package.json and tsconfig.json files for managing dependencies and TypeScript compiler options respectively
- These templates are designed to help set up a new project quickly with sensible defaults.

* Updated Pulumi template source and added new flags

The Makefile has been updated to change the source of the Pulumi templates from a GitHub URL to a local directory. Additionally, new flags have been introduced for name, description, and runtime options during the creation of apps. This allows for more flexibility and customization in app creation process.

* Updated Node version

The Node version specified in the .nvmrc file has been updated from 20.17.0 to 22.9.0 for better compatibility and performance improvements.