Refactor fs package and add new tests
The 'fs' package has been refactored to be part of the 'thecluster' package. The context functions have been updated to use the new 'thecluster.Fs' type instead of the old 'Fs'. A new test suite for the app package has also been added, along with a test for the Init function. The Init function itself has been modified to accept an instance of 'thecluster.Fs'. Lastly, some variable names in the Init function were changed for clarity.