Initial go at templating (#2910)
* Added context handling and initialization functionality

Introduced a new 'fs' package with context handling capabilities. This includes functions to retrieve and set filesystem instances from the context. Also, added an interface for the filesystem that uses 'afero.Fs'.

In addition, implemented an 'Init' function in the 'app' package which handles application initialization tasks such as retrieving git root directory, fetching templates from a repository, and error handling related to these operations.

* Added unit tests for fs context

Introduced a new set of unit tests to verify the functionality of the filesystem (fs) context. The test checks if the context can be successfully created and retrieved without any errors. It uses a mock filesystem provided by afero for testing purposes.

* Refactor fs package and add new tests

The 'fs' package has been refactored to be part of the 'thecluster' package. The context functions have been updated to use the new 'thecluster.Fs' type instead of the old 'Fs'. A new test suite for the app package has also been added, along with a test for the Init function. The Init function itself has been modified to accept an instance of 'thecluster.Fs'. Lastly, some variable names in the Init function were changed for clarity.

* Less terrible context stuff

* Reinventing the wheel

* cmd files

* Move the app command around

* Refactor template path and improve file handling

- Introduced a variable for the relative path to the Typescript templates, improving code readability.
- Modified the Init function to handle files more efficiently. It now locates the git root directory, creates relative paths, opens source files, creates destination files, and copies content from source to destination.
- Updated tests accordingly to reflect changes in file handling.