Introduced generic Result type and related functions
This update introduces a new generic Result type in the 'result' package. This type is a function that returns a value of any type and an error. Several methods and functions have been added to work with this new type, including Err, Value, IsErr, IsOk, Unwrap, Ok, Err (for creating Results), Error (for creating Results from string messages), Errorf (for formatted error messages), From (to create Results from values and errors) and zero (a helper function to get zero value of any type). Also included are Map and Bind functions for transforming Results.